Continuing the refactor, breaking up functionality which manages creating and updating channels from functionality which manages puzzle metadata.
