+++ bot/splat_store_cog.py
@@ -1,10 +1,12 @@
 import logging
+from typing import List, Optional
 
 import discord
 from discord.ext import commands 
 
 from bot import database
-from bot.database.models import HuntSettings 
+from bot.database.models import HuntSettings, PuzzleData
+from bot.data.puzzle_db import MissingPuzzleError, PuzzleDb
 
 logger = logging.getLogger(__name__)
 
@@ -42,3 +44,28 @@ class SplatStoreCog(commands.Cog):
         await interaction.response.send_message(f":exclamation: Most bot commands should be sent to #{settings.discord_bot_channel}")
         return False
 
+    async def get_puzzle_data_from_channel(self, channel) -> Optional[PuzzleData]:
+        """Extract puzzle data based on the channel name and category name
+
+        Looks up the corresponding JSON data
+        """
+        if not channel.category:
+            return None
+
+        guild = channel.guild
+        guild_id = guild.id
+        round_id = channel.category.id
+        round_name = channel.category.name
+        puzzle_id = channel.id
+        puzzle_name = channel.name
+        try:
+            return await PuzzleDb.get(guild_id, puzzle_id)
+        except MissingPuzzleError:
+            # Not the cleanest, just try to guess the original category id
+            # A DB would be useful here, then can directly query on solved_channel_id ..
+            logger.error(f"Unable to retrieve puzzle={puzzle_id} round={round_id} {round_name}/{puzzle_name}")
+            return None
+
+
+class GeneralAppError(RuntimeError):
+    pass
